% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagov.R
\name{cagov_resources}
\alias{cagov_resources}
\title{Get resources for a California dataset}
\usage{
cagov_resources(dataset_name)
}
\arguments{
\item{dataset_name}{Dataset name or ID}
}
\value{
tibble: id, name, format, url, description
}
\description{
Get resources for a California dataset
}
