% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagov.R
\name{cagov_search}
\alias{cagov_search}
\title{Search California Open Data datasets}
\usage{
cagov_search(query = NULL, rows = 50, start = 0)
}
\arguments{
\item{query}{Search term}

\item{rows}{Number of results (default 50)}

\item{start}{Offset (default 0)}
}
\value{
tibble: id, name, title, organization, num_resources, notes
}
\description{
Searches the CKAN catalog.
}
